first version
a handcrafted lexer is used to deal with the context-specific meaning
of the input characters, so the slightly modified ERE yacc parser code
published by opengroup can work as intended.

this version does nothing else than syntax check the passed ERE,
but it's already capable to parse very complex regexes.
the grammar in ere.y is considered bugfree, but there might still be
the one or other issue lingering in the lexer.