change prototype of matchfunc to take additional end pointer
since the length of the input expression might already be known
before calling the match function, having a gratuitous strlen()
call there will midly impact performance; additionally it requires
that string.h be included from the generated .c file (additional
to regex.h which is required for the regmatch_t declaration).
passing p and pe is in line with ragel idioms, so we use prefer
that over a startpointer+size "tuple".